[FLINK-3950] Fixed according to PR review
