+++ flink-metrics/flink-metrics-statsd/src/test/java/org/apache/flink/metrics/statsd/StatsDReporterTest.java
@@ -24,7 +24,7 @@ import org.apache.flink.configuration.Configuration;
 import org.apache.flink.metrics.Counter;
 import org.apache.flink.metrics.Histogram;
 import org.apache.flink.metrics.HistogramStatistics;
-import org.apache.flink.metrics.Meter;
+import org.apache.flink.metrics.util.TestMeter;
 import org.apache.flink.metrics.MetricConfig;
 import org.apache.flink.metrics.SimpleCounter;
 import org.apache.flink.metrics.reporter.MetricReporter;
@@ -226,7 +226,7 @@ public class StatsDReporterTest extends TestLogger {
 
 			registry = new MetricRegistry(config);
 			TaskManagerMetricGroup metricGroup = new TaskManagerMetricGroup(registry, "localhost", "tmId");
-			TestingMeter meter = new TestingMeter();
+			TestMeter meter = new TestMeter();
 			metricGroup.meter(meterName, meter);
 			String prefix = metricGroup.getMetricIdentifier(meterName);
 
@@ -324,40 +324,6 @@ public class StatsDReporterTest extends TestLogger {
 					return 6;
 				}
 			};
-		}
-	}
-
-	public static class TestingMeter implements Meter {
-
-		@Override
-		public void markEvent() { }
-
-		@Override
-		public void markEvent(long n) { }
-
-		@Override
-		public double getOneMinuteRate() {
-			return 1;
-		}
-
-		@Override
-		public double getFiveMinuteRate() {
-			return 5;
-		}
-
-		@Override
-		public double getFifteenMinuteRate() {
-			return 15;
-		}
-
-		@Override
-		public double getMeanRate() {
-			return 10;
-		}
-
-		@Override
-		public long getCount() {
-			return 100;
 		}
 	}
